pyproject.toml
[project]
name = "info-extract"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chardet>=5.2.0",
    "duckdb>=1.4.1",
    "extract-msg",
    "fastparquet>=2024.11.0",
    "jieba",
    "langextract>=1.0.9",
    "markdownify>=1.2.0",
    "msg-parser[rtf]>=1.2.0",
    "openai>=2.6.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.3",
    "pydantic-ai>=1.6.0",
    "python-dotenv>=1.1.1",
    "regex>=2025.10.23",
    "xlrd>=2.0.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
info-extract = "info_extract:async_main"
